Split pulsebuilding tool out to tool module
